pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digital-curator"
version = "0.1.0"
description = "Local-first photo triage application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "pillow>=10.0",
    "opencv-python>=4.9",
    "imagehash>=4.3",
    "send2trash>=1.8",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["host*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
